Simplify by combining types
Some types wrapped NSLayoutManager, NSTextStorage, and NSTextContainer so they could safely be exposed as public (without leaking implementation details). Since those types are no longer public, we don't need to wrap them. Therefore, combine where possible.
