+++ Sources/NativeMarkKit/render/LayoutManager.swift
@@ -7,7 +7,27 @@ import AppKit
 #error("Unsupported platform")
 #endif
 
+protocol LayoutManagerDelegate: AnyObject {
+    func layoutManager(_ layoutManager: LayoutManager, invalidateFrame frame: CGRect, inContainer container: TextContainer)
+}
+
 final class LayoutManager: NSLayoutManager {
+    weak var invalidationDelegate: LayoutManagerDelegate?
+    
+    override init() {
+        super.init()
+        delegate = self
+    }
+    
+    required init?(coder: NSCoder) {
+        super.init(coder: coder)
+        delegate = self
+    }
+    
+    var storage: NSTextStorage? {
+        textStorage
+    }
+    
     override func drawBackground(forGlyphRange glyphsToShow: NSRange, at origin: CGPoint) {
         super.drawBackground(forGlyphRange: glyphsToShow, at: origin)
         
@@ -15,9 +35,124 @@ final class LayoutManager: NSLayoutManager {
         drawBackgroundBorders(forGlyphRange: glyphsToShow)
         drawInlineBackgrounds(forGlyphRange: glyphsToShow)
     }
+    
+    func drawBackground(in bounds: CGRect, using styleSheet: StyleSheet) {
+        var attributes = [NSAttributedString.Key: Any]()
+        styleSheet.styles(for: .document).updateAttributes(&attributes)
+        if let backgroundColor = attributes[.backgroundColor] as? NativeColor {
+            backgroundColor.set()
+            bounds.fill()
+        } else {
+            bounds.clear()
+        }
+    }
+    
+    func accessibilityFrame(for characterRange: NSRange) -> CGRect {
+        let glyphRange = glyphRange(forCharacterRange: characterRange, actualCharacterRange: nil)
+        guard let container = textContainer(forGlyphAt: glyphRange.location, effectiveRange: nil) as? TextContainer else {
+            return .zero
+        }
+        
+        var rangeBounds = boundingRect(forGlyphRange: glyphRange, in: container)
+        enumerateEnclosingRects(forGlyphRange: glyphRange,
+                                withinSelectedGlyphRange: NSRange(location: NSNotFound, length: 0),
+                                in: container) { rect, stop in
+            rangeBounds = rect
+            stop.pointee = true
+        }
+        return rangeBounds + container.origin
+    }
+    
+    
+    func invalidateImage(in characterRange: NSRange) {
+        var actualRange = NSRange()
+        invalidateLayout(forCharacterRange: characterRange, actualCharacterRange: &actualRange)
+        invalidateDisplay(forCharacterRange: actualRange)
+        
+        let glyphRange = glyphRange(forCharacterRange: actualRange, actualCharacterRange: nil)
+        
+        if let container = textContainer(forGlyphAt: glyphRange.location, effectiveRange: nil) as? TextContainer {
+            let bounds = boundingRect(forGlyphRange: glyphRange, in: container)
+            invalidationDelegate?.layoutManager(self, invalidateFrame: bounds, inContainer: container)
+        } else {
+            // In this case, the image is big enough to force things out of the container
+            //  Brute force the updates
+            for wrappedContainer in textContainers {
+                let glyphRange = self.glyphRange(for: wrappedContainer)
+                let bounds = boundingRect(forGlyphRange: glyphRange, in: wrappedContainer)
+                if let container = wrappedContainer as? TextContainer {
+                    invalidationDelegate?.layoutManager(self, invalidateFrame: bounds, inContainer: container)
+                }
+            }
+        }
+    }
+    
+}
+
+extension LayoutManager: NSLayoutManagerDelegate {
+    func layoutManager(_ layoutManager: NSLayoutManager, shouldSetLineFragmentRect lineFragmentRect: UnsafeMutablePointer<CGRect>, lineFragmentUsedRect: UnsafeMutablePointer<CGRect>, baselineOffset: UnsafeMutablePointer<CGFloat>, in textContainer: NSTextContainer, forGlyphRange glyphRange: NSRange) -> Bool {
+        let characterRange = layoutManager.characterRange(forGlyphRange: glyphRange, actualGlyphRange: nil)
+        
+        var effectiveRange = characterRange
+        if let blockBackground = storage?.safeAttribute(.blockBackground, at: characterRange.location, effectiveRange: &effectiveRange) as? BackgroundValue {
+            let isAtStart = characterRange.location == effectiveRange.location
+            let isAtEnd = characterRange.upperBound == effectiveRange.upperBound
+            let defaultFont = storage?.safeAttribute(.font, at: characterRange.location, effectiveRange: nil) as? NativeFont ?? TextStyle.body.makeFont()
+            
+            let topMargin = isAtStart ? blockBackground.topMargin.asRawPoints(for: defaultFont.pointSize) : 0
+            let bottomMargin = isAtEnd ? blockBackground.bottomMargin.asRawPoints(for: defaultFont.pointSize) : 0
+            
+            let originalLineFragmentRect = lineFragmentRect.pointee
+            let updatedLineFragmentRect = CGRect(x: originalLineFragmentRect.minX,
+                                                 y: originalLineFragmentRect.minY,
+                                                 width: originalLineFragmentRect.width,
+                                                 height: originalLineFragmentRect.height + topMargin + bottomMargin)
+            lineFragmentRect.assign(repeating: updatedLineFragmentRect, count: 1)
+            
+            let originalUsedRect = lineFragmentUsedRect.pointee
+            let updatedUsedRect = CGRect(x: originalUsedRect.minX,
+                                         y: updatedLineFragmentRect.minY + topMargin,
+                                         width: originalUsedRect.width,
+                                         height: originalUsedRect.height)
+            lineFragmentUsedRect.assign(repeating: updatedUsedRect, count: 1)
+            
+            let originalBaseline = baselineOffset.pointee
+            baselineOffset.assign(repeating: originalBaseline + topMargin, count: 1)
+            
+            if isAtStart || isAtEnd {
+                return true
+            } // else fall through
+        }
+        
+#if false // DEBUG
+        if let storage = storage, let stringRange = Range(characterRange, in: storage.string) {
+            let string = storage.string[stringRange]
+            print("laying out [\(characterRange)](\(string)) at \(lineFragmentRect.pointee)")
+        }
+#endif
+        
+        return false
+    }
+    
+    func layoutManager(_ layoutManager: NSLayoutManager, shouldUse action: NSLayoutManager.ControlCharacterAction, forControlCharacterAt charIndex: Int) -> NSLayoutManager.ControlCharacterAction {
+        let ch = storage?.character(at: charIndex)
+        if ch == 13 /* \r */ {
+            return .lineBreak
+        } else if action.contains(.containerBreak),
+                  let containerBreakValue = storage?.safeAttribute(.containerBreak, at: charIndex, effectiveRange: nil) as? ContainerBreakValue {
+            return containerBreakValue.shouldContainerBreak ? .containerBreak : .zeroAdvancement
+        } else if action.contains(.lineBreak) {
+            // If we're about to actually break to a new container, don't also do a line break
+            let isAboutToContainerBreak = areAnyOfNextCharactersContainerBreaks(startingAt: charIndex + 1)
+            return isAboutToContainerBreak ? .zeroAdvancement : action
+        } else {
+            return action
+        }
+    }
 }
 
 private extension LayoutManager {
+    
     func drawBlockBackgrounds(forGlyphRange glyphsToShow: NSRange) {
         guard let storage = textStorage else {
             return
@@ -36,10 +171,10 @@ private extension LayoutManager {
             return
         }
         let frame = boundingRect(forGlyphRange: glyphRange, in: container) + container.origin
-        let defaultFont = textStorage?.attribute(.font, at: characterRange.location, effectiveRange: nil) as? NativeFont ?? TextStyle.body.makeFont()
+        let defaultFont = textStorage?.safeAttribute(.font, at: characterRange.location, effectiveRange: nil) as? NativeFont ?? TextStyle.body.makeFont()
         blockBackground.render(in: frame, defaultFont: defaultFont)
     }
-
+    
     func drawBackgroundBorders(forGlyphRange glyphsToShow: NSRange) {
         guard let storage = textStorage else {
             return
@@ -74,8 +209,8 @@ private extension LayoutManager {
     }
     
     func drawInlineBackground(_ inlineBackground: BackgroundValue, forCharacterRange characterRange: NSRange) {
-        let defaultFont = textStorage?.attribute(.font, at: characterRange.location, effectiveRange: nil) as? NativeFont ?? TextStyle.body.makeFont()
-
+        let defaultFont = textStorage?.safeAttribute(.font, at: characterRange.location, effectiveRange: nil) as? NativeFont ?? TextStyle.body.makeFont()
+        
         enumerateTypographicBounds(forCharacterRange: characterRange) { glyphRange, lineFrame, container in
             guard let textContainer = container as? TextContainer else {
                 return
@@ -86,7 +221,7 @@ private extension LayoutManager {
     
     func enumerateTypographicBounds(forCharacterRange characterRange: NSRange, with block: @escaping (NSRange, CGRect, NSTextContainer) -> Void) {
         let glyphRange = self.glyphRange(forCharacterRange: characterRange, actualCharacterRange: nil)
-
+        
         enumerateLineFragments(forGlyphRange: glyphRange) { lineFrame, usedRect, container, lineRange, _ in
             guard let intersectingRange = glyphRange.intersection(lineRange) else { return }
             
@@ -102,7 +237,7 @@ private extension LayoutManager {
         let glyphs = UnsafeMutablePointer<CGGlyph>.allocate(capacity: glyphRange.length)
         let characterIndices = UnsafeMutablePointer<Int>.allocate(capacity: glyphRange.length)
         let count = self.getGlyphs(in: glyphRange, glyphs: glyphs, properties: nil, characterIndexes: characterIndices, bidiLevels: nil)
-
+        
         var totalFrame: CGRect?
         for i in 0..<count {
             let glyph = glyphs[i]
@@ -117,17 +252,17 @@ private extension LayoutManager {
                 totalFrame = frame
             }
         }
-                
+        
         glyphs.deallocate()
         characterIndices.deallocate()
-
+        
         return totalFrame
     }
     
     func typographicBounds(ofGlyph glyph: CGGlyph, atGlyphIndex glyphIndex: Int, characterIndex: Int, onLineFragment lineFrame: CGRect, container: NSTextContainer) -> CGRect {
         let location = self.location(forGlyphAt: glyphIndex)
-        let defaultFont = self.textStorage?.attribute(.font, at: characterIndex, effectiveRange: nil) as? NativeFont ?? TextStyle.body.makeFont()
-
+        let defaultFont = self.textStorage?.safeAttribute(.font, at: characterIndex, effectiveRange: nil) as? NativeFont ?? TextStyle.body.makeFont()
+        
         let frameLocation = CGPoint(x: lineFrame.minX + location.x,
                                     y: lineFrame.minY + location.y)
         let glyphBoundingRect = boundingRect(forGlyphRange: NSRange(location: glyphIndex, length: 1),
@@ -138,5 +273,39 @@ private extension LayoutManager {
                       y: frameLocation.y - ascender,
                       width: glyphBoundingRect.width,
                       height: ascender - descender)
+    }
+    
+    func areAnyOfNextCharactersContainerBreaks(startingAt charIndex: Int) -> Bool {
+        var current = charIndex
+        var result = isContainerBreak(at: current)
+        while result != .notContainerBreak {
+            switch result {
+            case .notContainerBreak:
+                return false // stop entire process
+            case .isContainerBreakButDoesNotBreak:
+                break // go to the next character
+            case .isContainerBreak:
+                return true // we have one, so stop process
+            }
+            current += 1
+            result = isContainerBreak(at: current)
+        }
+        
+        return false
+    }
+    
+    enum Break {
+        case notContainerBreak
+        case isContainerBreakButDoesNotBreak
+        case isContainerBreak
+    }
+    
+    func isContainerBreak(at charIndex: Int) -> Break {
+        let length = storage?.length ?? 0
+        if charIndex < length, let containerBreakValue = storage?.safeAttribute(.containerBreak, at: charIndex, effectiveRange: nil) as? ContainerBreakValue {
+            return containerBreakValue.shouldContainerBreak ? .isContainerBreak : .isContainerBreakButDoesNotBreak
+        } else {
+            return .notContainerBreak
+        }
     }
 }
